renamed_package_to_org.mozilla.firefox.firefoxlitedownloader
