+++ FirefoxLite_Downloader_Library/src/main/java/org/mozilla/firefoxlitedownloaderlibrary/DownloadFileWorker.java
@@ -1,8 +1,7 @@
-package com.example.firefoxlitedownloaderlibrary;
+package org.mozilla.firefoxlitedownloaderlibrary;
 
 import android.content.Context;
 import android.os.Environment;
-import android.util.Log;
 
 import java.io.File;
 import java.io.FileOutputStream;
