+++ FirefoxLite_Downloader_Library/src/main/java/org/mozilla/firefoxlitedownloaderlibrary/LiveDataHelper.java
@@ -1,4 +1,4 @@
-package com.example.firefoxlitedownloaderlibrary;
+package org.mozilla.firefoxlitedownloaderlibrary;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MediatorLiveData;
