+++ FirefoxLite_Downloader_Library/src/main/java/org/mozilla/firefoxlitedownloaderlibrary/MainFunctionAccessFile.java
@@ -1,4 +1,4 @@
-package com.example.firefoxlitedownloaderlibrary;
+package org.mozilla.firefoxlitedownloaderlibrary;
 
 import java.util.concurrent.TimeUnit;
