+++ FirefoxLite_Downloader_Library/src/test/java/org/mozilla/firefoxlitedownloaderlibrary/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.firefoxlitedownloaderlibrary;
+package org.mozilla.firefoxlitedownloaderlibrary;
 
 import org.junit.Test;
