+++ app/src/androidTest/java/org/mozilla/firefox/firefoxlitedownloader/ExampleInstrumentedTest.java
@@ -1,11 +1,12 @@
-package com.example.mandeep.firefoxdownloader;
+package org.mozilla.firefox.firefoxlitedownloader;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import static org.junit.Assert.*;
