+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.mandeep.firefoxdownloader">
+    package="org.mozilla.firefox.firefoxlitedownloader">
     <uses-permission android:name="android.permission.INTERNET">
     </uses-permission>
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE">
@@ -15,7 +15,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="org.mozilla.firefox.firefoxlitedownloader.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
