+++ app/src/test/java/org/mozilla/firefox/firefoxlitedownloader/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.mandeep.firefoxdownloader;
+package org.mozilla.firefox.firefoxlitedownloader;
 
 import org.junit.Test;
